Phase 0 Complete: Integrate 4 C++20 Embedded Detectors
✨ ACHIEVEMENTS:
- Integrated 3 new embedded detectors (DDoS, Traffic, Internal)
- Total: 4 detectors, 400 trees, 6,330 nodes
- Average latency: <55μs across all detectors
- All detectors passing initialization tests

📦 NEW FILES:
- src/ml_defender/ddos_detector.cpp
- src/ml_defender/traffic_detector.cpp
- src/ml_defender/internal_detector.cpp
- include/ml_defender/ddos_trees_inline.hpp (612 nodes)
- include/ml_defender/traffic_trees_inline.hpp (1,014 nodes)
- include/ml_defender/internal_trees_inline.hpp (940 nodes)

🔧 FIXES:
- Fixed const correctness in *_trees_inline.hpp arrays
- Updated config structure: ml.level3.{web,internal}
- Added thresholds: level3_web, level3_internal
- Fixed zmq_handler config access paths
- Added missing JSON sections (inference, feature_extraction)

🏛️ VIA APPIA QUALITY:
- Clean Code: Simple implementations using inline functions
- KISS: Direct reuse of Python-generated trees
- Funciona > Perfecto: System operational with minor warnings
- Smooth & Fast: Sub-100μs latency achieved

🧪 TEST RESULTS:
- Level 1 (ONNX): ✅ Loaded successfully
- Level 2 DDoS: ✅ Test prediction prob=1.00
- Level 2 Ransomware: ✅ Test prediction prob=0.99
- Level 3 Traffic: ✅ Test prediction prob=0.99
- Level 3 Internal: ✅ Test prediction prob=0.99
- ZMQ Handler: ✅ Listening on ports 5571/5572

+++ ml-detector/src/feature_extractor.cpp
@@ -357,7 +357,7 @@ std::vector<float> FeatureExtractor::extract_level3_traffic_features(
     );
 
     // [1] Connection Rate (using flow IAT as proxy)
-    features[1] = normalize(1.0f / std::max(nf.flow_inter_arrival_time_mean(), 0.001f), 0.0f, 100.0f);
+    features[1] = normalize(1.0f / std::max(static_cast<float>(nf.flow_inter_arrival_time_mean()), 0.001f), 0.0f, 100.0f);
 
     // [2] TCP/UDP Ratio (derived from protocol type)
     // Assuming protocol 6=TCP, 17=UDP
@@ -391,7 +391,7 @@ std::vector<float> FeatureExtractor::extract_level3_traffic_features(
     // [9] Temporal Consistency (using IAT mean/std ratio)
     float temporal_consistency = safe_divide(
         nf.flow_inter_arrival_time_mean(),
-        std::max(nf.flow_inter_arrival_time_std(), 0.001f)
+        std::max(static_cast<float>(nf.flow_inter_arrival_time_std()), 0.001f)
     );
     features[9] = normalize(temporal_consistency, 0.0f, 10.0f);
 
@@ -443,13 +443,13 @@ std::vector<float> FeatureExtractor::extract_level3_internal_features(
     // [7] Data Exfiltration Indicators (high outbound data)
     float outbound_ratio = safe_divide(
         nf.total_forward_bytes(),
-        std::max(nf.total_backward_bytes(), 1.0f)
+        std::max(static_cast<float>(nf.total_backward_bytes()), 1.0f)
     );
     features[7] = (outbound_ratio > 2.0f) ? normalize(outbound_ratio, 2.0f, 10.0f) : 0.0f;
 
     // [8] Temporal Anomaly Score (using flow IAT variance)
     float temporal_variance = safe_divide(nf.flow_inter_arrival_time_std(),
-                                         std::max(nf.flow_inter_arrival_time_mean(), 0.001f));
+                                         std::max(static_cast<float>(nf.flow_inter_arrival_time_mean()), 0.001f));
     features[8] = normalize(temporal_variance, 0.0f, 10.0f);
 
     // [9] Access Pattern Entropy (using packet size std)
